[change] Agora o banco de dados é inicializado [change] Rotas de transações movidas para um novo REST
