+++ docker-compose-dev.yml
@@ -95,6 +95,8 @@ services:
       - account
     container_name: bantads-account-db
     image: 'postgres:13.1-alpine'
+    volumes:
+      - ./bds/account/init.sql:/docker-entrypoint-initdb.d/init.sql
     environment:
       - POSTGRES_PASSWORD=postgres
       - POSTGRES_USER=postgres
